move jblas_common to layers
