+++ intel_extension_for_transformers/llm/runtime/graph/models/model_utils/model_utils.cpp
@@ -42,8 +42,7 @@
 
 #include "core/ne_layers.h"
 #include "application/common.h"
-#include "core/jblas_common.hpp"
-#include "jblas/jit_blas_weight_compression.h"
+#include "core/layers/jblas_common.hpp"
 #include "models/model_utils/model_files.h"
 #include "models/model_utils/model_utils.h"
 #include "models/model_utils/util.h"
